Migrate tests from FunSuite to FunSpec style
- Replace all *Suite.scala test files with *Spec.scala equivalents
- Update test structure to use describe/it pattern instead of test()
- Replace shouldEqual assertions with standard assert() calls
