+++ arktwin/edge/src/test/scala/arktwin/edge/test/ActorTestBase.scala
@@ -4,8 +4,8 @@ package arktwin.edge.test
 
 import org.apache.pekko.actor.testkit.typed.scaladsl.ScalaTestWithActorTestKit
 import org.scalatest.BeforeAndAfterAll
-import org.scalatest.funsuite.AnyFunSuiteLike
+import org.scalatest.funspec.AnyFunSpecLike
 
-trait ActorTestBase extends ScalaTestWithActorTestKit with AnyFunSuiteLike with BeforeAndAfterAll:
+trait ActorTestBase extends ScalaTestWithActorTestKit with AnyFunSpecLike with BeforeAndAfterAll:
   override def beforeAll() =
     scribe.Logger.root.withMinimumLevel(scribe.Level.Error).replace()
